+++ android/app/src/main/java/com/yagubogu/ui/common/component/chart/AnimatedBarChart.kt
@@ -1,4 +1,4 @@
-package com.yagubogu.ui.common.component
+package com.yagubogu.ui.common.component.chart
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
