+++ android/app/src/main/java/com/yagubogu/ui/stats/detail/component/StadiumVisitCounts.kt
@@ -12,7 +12,7 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.yagubogu.R
-import com.yagubogu.ui.common.component.AnimatedBarChart
+import com.yagubogu.ui.common.component.chart.AnimatedBarChart
 import com.yagubogu.ui.common.model.BarChartItemValue
 import com.yagubogu.ui.common.model.BarChartLabel.Companion.DefaultBarChartDataLabel
 import com.yagubogu.ui.common.model.BarChartLabel.Companion.DefaultBarChartTitleLabel
